# test(ivy): clean up class_binding perf tests for readability