New reporting format:
```

Benchmark: class binding[create]: <div class="A B">
 styling: 287.289 ns(-44%) 
 base: 199.771 ns(0%) 

Benchmark: class binding[update]: <div class="A B">
 styling: 62.092 us(-1%) 
 base: 61.202 us(0%) 

Benchmark: class binding[noop]: <div class="A B">
 styling: 62.187 us(-2%) 
 base: 61.100 us(0%) 

Benchmark: class binding[create]: <div [class]="toggle ? A_1 : B_1">
 styling: 199.370 ns(-7%) 
 base: 185.834 ns(0%) 

Benchmark: class binding[update]: <div [class]="toggle ? A_1 : B_1">
 styling: 149.971 us(-79%) 
 base: 83.769 us(0%) 

Benchmark: class binding[noop]: <div [class]="toggle ? A_1 : B_1">
 styling: 115.710 us(-73%) 
 base: 66.950 us(0%) 

Benchmark: class binding[create]: <div [class]="toggle ? A_10 : B_10">
 styling: 199.028 ns(-7%) 
 base: 185.148 ns(0%) 

Benchmark: class binding[update]: <div [class]="toggle ? A_10 : B_10">
 styling: 141.881 us(-67%) 
 base: 84.752 us(0%) 

Benchmark: class binding[noop]: <div [class]="toggle ? A_10 : B_10">
 styling: 108.357 us(-62%) 
 base: 66.933 us(0%) 

Benchmark: class binding[create]: <div [class]="toggle ? A_1 : B_1">
 styling: 198.919 ns(-7%) 
 base: 186.578 ns(0%) 

Benchmark: class binding[update]: <div [class]="toggle ? A_1 : B_1">
 styling: 141.638 us(-67%) 
 base: 84.988 us(0%) 

Benchmark: class binding[noop]: <div [class]="toggle ? A_1 : B_1">
 styling: 109.980 us(-63%) 
 base: 67.617 us(0%) 

Benchmark: class binding[create]: <div class="A B" [class]="toggle ? A_1 : B_1">
 styling: 291.815 ns(-56%) 
 base: 187.385 ns(0%) 

Benchmark: class binding[update]: <div class="A B" [class]="toggle ? A_1 : B_1">
 styling: 980.686 us(-1044%) 
 base: 85.712 us(0%) 

Benchmark: class binding[noop]: <div class="A B" [class]="toggle ? A_1 : B_1">
 styling: 118.095 us(-70%) 
 base: 69.311 us(0%) 

Benchmark: class binding[create]: <div class="A B" [class]="toggle ? A_10 : B_10">
 styling: 257.927 ns(-47%) 
 base: 175.274 ns(0%) 

Benchmark: class binding[update]: <div class="A B" [class]="toggle ? A_10 : B_10">
 styling: 21044.299 us(-24776%) 
 base: 84.598 us(0%) 

Benchmark: class binding[noop]: <div class="A B" [class]="toggle ? A_10 : B_10">
 styling: 117.065 us(-66%) 
 base: 70.537 us(0%) 

Benchmark: class binding[create]: <div class="A B" [class]="toggle ? A_1 : B_1" [class.foo]="toggle">
 styling: 257.294 ns(-44%) 
 base: 178.327 ns(0%) 

Benchmark: class binding[update]: <div class="A B" [class]="toggle ? A_1 : B_1" [class.foo]="toggle">
 styling: 1090.994 us(-1089%) 
 base: 91.774 us(0%) 

Benchmark: class binding[noop]: <div class="A B" [class]="toggle ? A_1 : B_1" [class.foo]="toggle">
 styling: 146.925 us(-107%) 
 base: 71.125 us(0%) 
```